+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added new `extraConfirmText`, `extraConfirmLabel` properties to `MessageOptions`.  Use this option
   to require the specified text to be re-typed by a user when confirming a potentially destructive or disruptive action.
+* Updated grid column filters to apply on `Enter` / dismiss on `Esc` and tweaked the filter popup toolbar for clarity.
 
 ## 75.0.1 - 2025-08-11